feat: change AddCard to hooks
