+++ src/components/Search/Search.tsx
@@ -1,4 +1,4 @@
-import React, { FC, useEffect, useState } from 'react';
+import { useEffect, useState } from 'react';
 import styles from './Search.module.css';
 
 const Search = () => {
